Deduplicate tensor-splitting utility (#5545)
